+++ src/lib/utils.ts
@@ -76,3 +76,25 @@ export function convertTime(duration: string) {
     }`.trim() || "0 min"
   );
 }
+
+export function isSectionedInstruction(
+  instructions: Instructions
+): instructions is SectionedInstructions {
+  return (
+    Array.isArray(instructions) &&
+    instructions.length > 0 &&
+    typeof instructions[0] === "object" &&
+    instructions[0] !== null &&
+    "name" in instructions[0] &&
+    "text" in instructions[0]
+  );
+}
+
+export function isSimpleInstruction(
+  instructions: Instructions
+): instructions is SimpleInstructions {
+  return (
+    Array.isArray(instructions) &&
+    (instructions.length === 0 || typeof instructions[0] === "string")
+  );
+}
